NewsPaper/requirements.txt
Django==4.2.2      
django-filter==23.2
django-allauth